Speed Booster code started, room code cleanup, minor stuff
Main thing in this update is the start of the basic Speed Booster code. I will include more info on this in a blog post, but it will be different from most Metroid games, and be more of a Hollow Knight-style dashing ability. It also will not be permanently named the Speed Booster.

Added some room information to scripts that I forgot. I apparently need to make a note to myself when I add new rooms which scripts I need to update.

Changed syntax in some input code. Edited player mask to investigate a possible minor bug in smooth collisions. Will figure it out at a later date, but it is worth noting that it may be needed to separate movement from all other player actions with objects in this environment. More on that when it becomes a real problem.