Arc Dash audio assets, started HUD rebuild side-project, etc.
Added more Arc Dash assets, mostly audio. Added 8x8 and 16x16 blocks to Arc Dash destructibles list. Started an audio initialization script. Adding only new sounds as I go.

Started a side-project to redraw HUD elements. The purpose will be to incorporate the map system to be square-shaped instead of a skewed rectangle. This will allow us to implement the feature sooner. This is not a high-priority project, so it will be something to work on when I am not actively working on the demo.